auth: simplify public auth exception surface
